Fix for segmentation fault when running wave_stat.x in the global workflow (#102)
This PR changes single to double precision in wave_stat.f90 to fix segmentation
faults. Credit to @AminIlia-NOAA for pointing out the same issue he had
with oceanicepost job and how it was fixed there.

Resolves #98

+++ src/wave_stat.fd/wave_stat.f90
@@ -29,7 +29,7 @@
       integer :: ymdc,fhr,nnme,nnsc
       integer :: parcode(3)
       integer :: istop
-      real,allocatable :: dumy1d(:)
+      real(8),allocatable :: dumy1d(:)
 ! Grid
       real    :: rdlon(3),rdlat(3)
       integer :: nlola(2)
